refactor(plugin-contentful): split manager contentful class
