+++ packages/botonic-plugin-contentful/src/contentful/manage/manage-entry.ts
@@ -1,0 +1,174 @@
+// eslint-disable-next-line node/no-missing-import
+import { ClientAPI } from 'contentful-management/dist/typings/create-contentful-api'
+// eslint-disable-next-line node/no-missing-import
+import { Entry } from 'contentful-management/dist/typings/entities/entry'
+// eslint-disable-next-line node/no-missing-import
+import { Environment } from 'contentful-management/dist/typings/entities/environment'
+
+import { CmsException, ContentId, ContentType } from '../../cms'
+import { ResourceNotFoundCmsException } from '../../cms/exceptions'
+import {
+  CONTENT_FIELDS,
+  ContentField,
+  ContentFieldType,
+} from '../../manage-cms/fields'
+import { ManageContext } from '../../manage-cms/manage-context'
+import * as nlp from '../../nlp'
+import { ContentfulOptions } from '../../plugin'
+import { isOfType } from '../../util/enums'
+
+export class ManageContentfulEntry {
+  constructor(
+    readonly options: ContentfulOptions,
+    readonly manage: ClientAPI,
+    readonly environment: Promise<Environment>
+  ) {}
+
+  async updateFields(
+    context: ManageContext,
+    contentId: ContentId,
+    fields: { [contentFieldType: string]: any }
+  ): Promise<{ [contentFieldType: string]: any }> {
+    const environment = await this.environment
+    const oldEntry = await this.getEntry(environment, contentId)
+    let needUpdate = false
+    for (const key of Object.keys(fields)) {
+      if (!isOfType(key, ContentFieldType)) {
+        throw new CmsException(`'${key}' is not a valid content field type`)
+      }
+      const field = this.checkOverwrite(context, oldEntry, key, false)
+      if (oldEntry.fields[field.cmsName][context.locale] === fields[key]) {
+        continue
+      }
+      needUpdate = true
+      oldEntry.fields[field.cmsName][context.locale] = fields[key]
+    }
+    if (!needUpdate) {
+      return oldEntry.fields
+    }
+    // we could use this.deliver.contentFromEntry & IgnoreFallbackDecorator to convert
+    // the multilocale fields returned by update()
+    await this.writeEntry(context, oldEntry)
+    return oldEntry.fields
+  }
+
+  async createContent(
+    context: ManageContext,
+    model: ContentType,
+    id: string
+  ): Promise<void> {
+    const environment = await this.environment
+    try {
+      await environment.createEntryWithId(model, id, {
+        fields: {},
+      })
+    } catch (e) {
+      throw new CmsException('ERROR while creating content', e)
+    }
+  }
+
+  async deleteContent(
+    context: ManageContext,
+    contentId: ContentId
+  ): Promise<void> {
+    const environment = await this.environment
+    const oldEntry = await this.getEntry(environment, contentId)
+    if (oldEntry) {
+      await oldEntry.unpublish()
+      await oldEntry.delete()
+    }
+  }
+
+  private checkOverwrite(
+    context: ManageContext,
+    entry: Entry,
+    fieldType: ContentFieldType,
+    failIfMissing: boolean
+  ): ContentField {
+    if (entry.isArchived()) {
+      throw new CmsException(`Cannot update an archived entry`)
+    }
+    const field = CONTENT_FIELDS.get(fieldType)
+    if (!field) {
+      throw new CmsException(`Invalid field type ${fieldType}`)
+    }
+    if (!context.locale) {
+      // paranoic check
+      throw new Error('Context.locale must be defined')
+    }
+    if (!(field.cmsName in entry.fields)) {
+      if (!failIfMissing) {
+        entry.fields[field.cmsName] = {}
+        return field
+      }
+      const fields = Object.keys(entry.fields)
+      throw new CmsException(
+        `Field '${field.cmsName}' not found in entry of type '${
+          entry.sys.contentType.sys.id
+        }. It only has ${JSON.stringify(fields)}'`
+      )
+    }
+    if (!context.allowOverwrites) {
+      const value = entry.fields[field.cmsName][context.locale]
+      if (value) {
+        throw new CmsException(
+          `Cannot overwrite field '${field.cmsName}' of entry '${
+            entry.sys.id
+          }' "+
+          "(has value '${String(
+            value
+          )}') because ManageContext.allowOverwrites is false`
+        )
+      }
+    }
+    return field
+  }
+
+  async copyField(
+    context: ManageContext,
+    contentId: ContentId,
+    fieldType: ContentFieldType,
+    fromLocale: nlp.Locale,
+    onlyIfTargetEmpty: boolean
+  ): Promise<void> {
+    const environment = await this.environment
+    const oldEntry = await environment.getEntry(contentId.id)
+    const field = this.checkOverwrite(context, oldEntry, fieldType, false)
+
+    const fieldEntry = oldEntry.fields[field.cmsName]
+    if (fieldEntry == undefined) {
+      return
+    }
+    // TODO shouldn't this check be done before checkOverwrite?
+    if (onlyIfTargetEmpty && context.locale in fieldEntry) {
+      return
+    }
+    fieldEntry[context.locale] = fieldEntry[fromLocale]
+    await this.writeEntry(context, oldEntry)
+  }
+
+  private async writeEntry(
+    context: ManageContext,
+    entry: Entry
+  ): Promise<void> {
+    if (context.dryRun) {
+      console.log('Not updating due to dryRun mode')
+      return
+    }
+    const updated = await entry.update()
+    if (!context.preview) {
+      await updated.publish()
+    }
+  }
+
+  async getEntry(
+    environment: Environment,
+    contentId: ContentId
+  ): Promise<Entry> {
+    try {
+      return await environment.getEntry(contentId.id)
+    } catch (e) {
+      throw new ResourceNotFoundCmsException(contentId, e)
+    }
+  }
+}
